+++ .gitignore
@@ -12,5 +12,3 @@ markdown.py
 examples
 tests
 rid.egg-info
-
-
