+++ rid/enhcMD.py
@@ -5,16 +5,26 @@ from rid.lib.utils import replace
 from rid.lib.utils import make_iter_name
 from rid.lib.utils import make_walker_name
 from rid.lib.utils import checkfile
-from rid.lib.utils import log_task
+from ridkit.lib.utils import cmd_append_log, set_resource, set_machine, log_task
+
 from rid.lib.gen.gen_mdp import make_grompp
 from rid.lib.gen.gen_plumed import make_plumed
 from rid.lib.gen.gen_plumed import conf_enhc_plumed
+from ridkit.lib.cal_cv_dim import cal_cv_dim
+from ridkit.lib.cmpf import cmpf
+
+from dpdispatcher.submission import Submission, Job, Task, Resources
+from dpdispatcher.lazy_local_context import LazyLocalContext
 
 enhc_name="00.enhcMD"
-
 enhc_plm="plumed.dat"
 enhc_bf_plm="plumed.bf.dat"
 enhc_out_plm="plm.out"
+enhc_out_conf="confs/"
+enhc_out_angle="angle.rad.out"
+
+res_name="01.resMD"
+res_plm="plumed.res.dat"
 
 def adjust_lvl(prev_enhc_path, num_of_cluster_threshhold, jdata):
     # adaptive trust level
@@ -133,3 +143,97 @@ def make_enhc (iter_index,
             log_task ("set trust l1 and l2: %f %f" % (enhc_trust_lvl_1, enhc_trust_lvl_2))
     print("Enhanced sampling has prepared.")
 
+
+def post_enhc (iter_index, 
+               json_file,
+               base_dir="./") :
+    base_dir = os.path.abspath(base_dir) + "/"
+    iter_name = make_iter_name (iter_index)
+    work_path = base_dir + iter_name + "/" + enhc_name + "/" 
+    json_file = os.path.abspath(json_file) 
+    json_file = os.path.abspath(json_file)
+    fp = open (json_file, 'r')
+    jdata = json.load (fp)
+    fp.close()
+    gmx_split = jdata["gmx_split_traj"]
+    gmx_split_log = "gmx_split.log"
+    gmx_split_cmd = cmd_append_log (gmx_split, gmx_split_log)
+    
+    all_task = list(filter(lambda x:os.path.isdir(x),  glob.glob(work_path + "/[0-9]*[0-9]")))
+    all_task.sort()
+
+    cwd = os.getcwd()
+    numb_walkers = jdata["numb_walkers"]
+    for ii in range(numb_walkers) :
+        walker_path = work_path + make_walker_name(ii) + "/"
+        os.chdir(walker_path)        
+        if os.path.isdir ("confs") : 
+            shutil.rmtree ("confs")
+        os.makedirs ("confs")
+        os.chdir(cwd)
+
+    print('rid.py:post_enhc:gmx_split_cmd', gmx_split_cmd)
+    print('rid.py:post_enhc:work path', work_path)
+
+    machine = set_machine(json_file)
+    resources = set_resource(json_file, target="post")
+    all_task_relpath = [os.path.relpath(ii, work_path) for ii in all_task]
+    gmx_split_task = [ Task(command=gmx_split_cmd, task_work_path=ii, outlog='gmx_split.log', errlog='gmx_split.log') for ii in all_task_relpath ]
+    gmx_split_submission = Submission(work_base=work_path, resources=resources, batch=machine, task_list=gmx_split_task)
+    gmx_split_submission.run_submission()
+    
+    for ii in range(numb_walkers) :
+        walker_path = work_path + make_walker_name(ii) + "/"
+        angles = np.loadtxt (walker_path + enhc_out_plm)
+        np.savetxt (walker_path + enhc_out_angle, angles[:,1:], fmt="%.6f")
+    print("Post process of enhanced sampling finished.")
+
+
+def run_enhc (iter_index,
+              json_file,
+              base_dir='./') :
+    json_file = os.path.abspath(json_file)
+    base_dir = os.path.abspath(base_dir) + "/"
+    iter_name = make_iter_name (iter_index)
+    work_path = base_dir + iter_name + "/" + enhc_name + "/"  
+
+    fp = open (json_file, 'r')
+    jdata = json.load (fp)
+    fp.close()
+    gmx_prep = jdata["gmx_prep"]
+    gmx_run = jdata["gmx_run"]
+    enhc_thread = jdata["enhc_thread"]
+    gmx_run = gmx_run + (" -nt %d" % enhc_thread)
+    gmx_prep_log = "gmx_grompp.log"
+    gmx_run_log = "gmx_mdrun.log"
+    # assuming at least one walker
+    graph_files = glob.glob (work_path + (make_walker_name(0)) + "/*.pb")
+    if len (graph_files) != 0 :
+        gmx_run = gmx_run + " -plumed " + enhc_plm  
+    else :
+        gmx_run = gmx_run + " -plumed " + enhc_bf_plm
+    gmx_prep_cmd = cmd_append_log (gmx_prep, gmx_prep_log)
+    gmx_run_cmd = cmd_append_log (gmx_run, gmx_run_log)
+    numb_walkers = jdata["numb_walkers"]
+
+    all_task = list(filter(lambda x:os.path.isdir(x),  glob.glob(work_path + "/[0-9]*[0-9]")))
+    all_task.sort()
+
+    all_task_basedir = [os.path.relpath(ii, work_path) for ii in all_task]
+    print('run_enhc:work_path', work_path)
+    print('run_enhc:gmx_prep_cmd:', gmx_prep_cmd)
+    print('run_enhc:gmx_run_cmd:', gmx_run_cmd)
+    print('run_enhc:all_task:', all_task)
+    print('run_enhc:all_task_basedir:', all_task_basedir)
+    
+    machine = set_machine(json_file)
+    resources = set_resource(json_file, target="enhc")
+
+    gmx_prep_task = [ Task(command=gmx_prep_cmd, task_work_path=ii, outlog='gmx_grompp.log', errlog='gmx_grompp.log') for ii in all_task_basedir ]
+    gmx_prep_submission = Submission(work_base=work_path, resources=resources, batch=machine, task_list=gmx_prep_task)
+
+    gmx_prep_submission.run_submission()
+    
+    gmx_run_task =  [ Task(command=gmx_run_cmd, task_work_path=ii, outlog='gmx_mdrun.log', errlog='gmx_mdrun.log') for ii in all_task_basedir ]
+    gmx_run_submission = Submission(work_base=work_path, resources=resources, batch=machine, task_list=gmx_run_task)
+    gmx_run_submission.run_submission()
